Extract implementation for status handling into a MicoStatusService class
